// assets/api/dto/ITeam.ts
export interface ITeam {
  country: string;
  flag: string;
  foundation_date: number;
  full_name: string;
  league_name: string;
  league_slug: string;
  league_uuid: string;
  logo: string;
  manager_name: string;
  meta_description: string;
  meta_keywords: string;
  meta_title: string;
  name: string;
  participant_descriptions: string;
  participant_alpha2: string;
  participant_country: string;
  participant_flag: string;
  slug: string;
  sport_name: string;
  sport_url: string;
  stadium_capacity: number;
  stadium_city_name: string;
  stadium_name: string;
  uuid: string;
}